cyber-brain/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.1.0
requests>=2.32.0
